+++ src/server/commands/generation/gen.ts
@@ -1,10 +1,8 @@
 import { registerCommand } from '../register_commands.js';
 import { ExpressionShape } from '../../shapes/expression.js';
-import { regionSize } from '../../util.js';
 import { assertCuboidSelection, assertCanBuildWithin } from '@modules/assert.js';
 import { Pattern } from '@modules/pattern.js';
-import { RawText } from '@notbeer-api';
-import { Vector } from '@notbeer-api';
+import { RawText, Vector, regionSize } from '@notbeer-api';
 
 const registerInformation = {
     name: 'gen',
